Organized cosmwasm/enclaves/shared/utils and a couple other small changes
