+++ cosmwasm/enclaves/execute/src/lib.rs
@@ -1,4 +1,3 @@
-#![feature(try_reserve)]
 // similar trick to get the IDE to use sgx_tstd even when it doesn't know we're targeting SGX
 #[cfg(not(target_env = "sgx"))]
 extern crate sgx_tstd as std;
