+++ cosmwasm/enclaves/shared/utils/src/results.rs
@@ -1,0 +1,29 @@
+use log::*;
+
+use sgx_types::*;
+
+pub trait UnwrapOrSgxErrorUnexpected {
+    type ReturnType;
+    fn sgx_error(self) -> Result<Self::ReturnType, sgx_status_t>;
+    fn sgx_error_with_log(self, err_mgs: &str) -> Result<Self::ReturnType, sgx_status_t>;
+}
+
+impl<T, S> UnwrapOrSgxErrorUnexpected for Result<T, S> {
+    type ReturnType = T;
+    fn sgx_error(self) -> Result<Self::ReturnType, sgx_status_t> {
+        match self {
+            Ok(r) => Ok(r),
+            Err(_) => Err(sgx_status_t::SGX_ERROR_UNEXPECTED),
+        }
+    }
+
+    fn sgx_error_with_log(self, log_msg: &str) -> Result<Self::ReturnType, sgx_status_t> {
+        match self {
+            Ok(r) => Ok(r),
+            Err(_) => {
+                error!("{}", log_msg);
+                Err(sgx_status_t::SGX_ERROR_UNEXPECTED)
+            }
+        }
+    }
+}
